requirements.txt
discord.py==2.3.2
gql==3.5.0
python-dotenv==1.0.1
openai==1.66.3
requests==2.31.0
requests-toolbelt==1.0.0
fastapi==0.110.0
uvicorn==0.27.1 